Widen flag argument of CReadUtil::AsSeqId() to silence compiler warnings (and guard against lurking bugs).
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@99183 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/objtools/readers/gff_base_columns.hpp
@@ -44,7 +44,7 @@ class CGffBaseColumns
 {
 public:
     using TFrame = CCdregion::EFrame;
-    using SeqIdResolver = CRef<CSeq_id> (*)(const string&, unsigned int, bool);
+    using SeqIdResolver = CRef<CSeq_id> (*)(const string&, long, bool);
 
     CGffBaseColumns();
